# services/gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateUserRequest:
    properties:
      email:
        example: manager@example.com
        type: string
      password:
        example: Passw0rd!123
        type: string
      username:
        example: manager01
        type: string
    type: object
...
  dto.UpdateUserActivityRequest:
    properties:
      is_active:
        example: true
        type: boolean
    type: object
  dto.UpdateUserActivityResponse:
    properties:
      is_active:
        example: true
        type: boolean
    type: object
  dto.UpdateUserRequest:
    properties:
      email:
        example: manager.updated@example.com
        type: string
      username:
        example: manager02
        type: string
    type: object
  dto.UpdateUserRoleRequest:
    properties:
      role:
        example: USER_ROLE_USER, USER_ROLE_MODERATOR, USER_ROLE_ADMIN
        type: string
    type: object
  dto.UpdateUserRoleResponse:
    properties:
      role:
        example: ROLE_ADMIN
        type: string
    type: object
  dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  dto.UsersResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
    type: object
  responder.ErrorResponse:
    properties:
      details:
        items:
          $ref: '#/definitions/responder.ErrorResponseDetails'
        type: array
      error:
        type: string
    type: object
  responder.ErrorResponseDetails:
    properties:
      description:
        type: string
      field:
        type: string
    type: object
info:
  contact: {}
  description: REST API Gateway for Fukuro Reserve microservices.
  title: Swagger Gateway API
  version: "1.0"
paths:
...
  /users:
    get:
      description: Returns users with pagination. Requires JWT auth.
      parameters:
      - default: 1
        description: Page number (starts from 1)
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Creates a user account (admin/staff operation). Requires JWT auth.
      parameters:
      - description: Create user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      description: Deletes user by ID. Requires JWT auth.
      parameters:
      - description: User ID
        example: 1
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete user
      tags:
      - users
    get:
      description: Returns a user by ID. Requires JWT auth.
      parameters:
      - description: User ID
        example: 1
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
      security:
      - Bearer: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Updates user profile fields. Requires JWT auth.
      parameters:
      - description: User ID
        example: 1
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Update user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
      security:
      - Bearer: []
      summary: Update user
      tags:
      - users
  /users/{id}/activity:
    patch:
      consumes:
      - application/json
      description: Updates active/inactive flag for a user. Requires JWT auth.
      parameters:
      - description: User ID
        example: 1
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Update activity request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserActivityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateUserActivityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
      security:
      - Bearer: []
      summary: Update user activity status
      tags:
      - users
  /users/{id}/role:
    patch:
      consumes:
      - application/json
      description: Updates user role. Requires JWT auth.
      parameters:
      - description: User ID
        example: 1
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Update role request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateUserRoleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
      security:
      - Bearer: []
      summary: Update user role
      tags:
      - users
  /users/me:
    get:
      description: Returns current user profile by access token. Requires JWT auth.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responder.ErrorResponse'
      security:
      - Bearer: []
      summary: Get current user
      tags:
      - users
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"